load rails in delayed
